Move testing framework into test_framework directory
