+++ contrib/highfidelity/test_framework/alice_and_bob.py
