+++ contrib/highfidelity/test_framework/blockchain/elements.bootstrap_signing_key.conf
